@@ -10,27 +10,29 @@
     "nwdev": "concurrently --kill-others --raw \"npm run dev\" \"./node_modules/.bin/run --mac --x64 ./launch\" "
   },
   "dependencies": {
+    "@vue/cli-service": "^3.0.0-rc.3",
     "nwjs-builder-phoenix": "^1.14.6",
     "vue": "^2.5.3",
     "vue-router": "^3.0.1",
-    "vuetify": "^1.0.0",
+    "vuetify": "^1.1.1",
     "vuex": "^3.0.1"
   },
   "devDependencies": {
     "babel-core": "^6.0.0",
-    "babel-loader": "^6.0.0",
+    "babel-loader": "^7.1.4",
+    "babel-plugin-transform-imports": "^1.4.1",
     "babel-preset-env": "^1.6.0",
     "babel-preset-stage-2": "^6.24.1",
-    "babel-plugin-transform-imports": "^1.4.1",
     "cross-env": "^3.0.0",
     "css-loader": "^0.25.0",
     "file-loader": "^0.9.0",
+    "prettier": "^1.13.7",
     "style-loader": "^0.13.1",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.1",
     "vue-loader": "^12.1.0",
     "vue-template-compiler": "^2.5.3",
-    "webpack": "^2.6.1",
+    "webpack": "^3.0.0",
     "webpack-dev-server": "^2.4.5"
   },
   "build": {
