@@ -1,7 +1,9 @@
 <template>
      <v-list-tile avatar>
          <v-list-tile-avatar>
-             <v-icon class="white--text">report_problem</v-icon>
+             <transition name="fade">
+                <v-icon class="white--text" :class="icon.classes">{{icon.name}}</v-icon>
+             </transition>
          </v-list-tile-avatar>
          <v-list-tile-content>
              <v-list-tile-title>{{remote.alias}}</v-list-tile-title>
@@ -30,7 +32,14 @@
 </template>
 
 <script>
+import Monitor from "../mixins/pingLogic";
 export default {
+  mixins: [Monitor],
+  data() {
+    return {
+      currentStatus: this.remote.status
+    };
+  },
   props: {
     remote: {
       type: Object,
@@ -50,9 +59,40 @@ export default {
     removeRemote(id) {
       this.$store.commit("deleteRemote", id);
     }
+  },
+  computed: {
+    icon() {
+      let name = "";
+      let classes = [];
+
+      if (!this.currentStatus || this.currentStatus === "-") {
+        name = "report_problem";
+        classes = ["grey", "darken-2"];
+      } else {
+        if (this.currentStatus === "online") {
+          name = "swap_vert";
+          classes = ["light-green", "darken-1"];
+        } else if (this.currentStatus === "offline") {
+          name = "not_interested";
+          classes = ["red", "darken-1"];
+        }
+      }
+      return {
+        name,
+        classes
+      };
+    }
   }
 };
 </script>
 
 <style>
+.fade-enter-active,
+.fade-leave-active {
+  transition: opacity 0.8s;
+}
+.fade-enter,
+.fade-leave-to {
+  opacity: 0;
+}
 </style>
